+++ 놀이터(예제 코드 작성)/jackson/src/main/kotlin/com/banjjoknim/playground/model/Car.kt
@@ -1,3 +1,3 @@
-package com.banjjoknim.playground.jsonserialize
+package com.banjjoknim.playground.model
 
 data class Car(val name: String, val price: Int = 0)
